Highlight special classes `self` and `parent`
Motivation:
Currently, `self` and `parent` are highlighted as `@variable.builtin`, but no
default theme [3] defines a highlighting for `@variable.builtin`, so they
end up highlighted as normal variables. This is in stark contrast with `static`
(which is highlighted as keyword) because they are used in similiar way.

Context:
* `self`, `parent` and `static` are "special classes" [1]
* `static` is also a "keyword" (but `self` and `parent` aren't) [2]
* `relative_scope` refers to `self`, `parent` and `static` [4]
* `static` highlighting is overridden by being a keyword

Solution:
* I changed `relative_scope` from `@variable.builtin` to `@contructor`,
because they can be used in the same place as a constructor (e.g.
`static::Foo()` and `MyClass::Foo()`)
* `relative_scope` doesn't hande `new self()` and `$x instanceof self`, so
I added rules for this two cases. Also as `@constructor` (to match
`new MyClass()` and `$x instanceof MyClass`)

Alternatives:
* Maybe the themes should be extended to highlight `@variable.builtin`, instead
of changing the php extension?
* Maybe they should be highlighted as keyword to match `static`?

Links:
* [1] https://www.php.net/manual/en/reserved.classes.php#reserved.classes.special
* [2] https://www.php.net/manual/en/reserved.keywords.php
* [3] https://github.com/zed-industries/zed/tree/main/assets/themes
* [4] https://github.com/tree-sitter/tree-sitter-php/blob/master/common/define-grammar.js#L1200